Improved fix 600
Added tests to test_raw.py
Changes ValueError to TypeError
